Subscription_status session variable
Added a new session variable to check subscription status. This can be used to show different elements/options on screen depending on whether the user is subscribed. I intend on using this variable to change the way checkout works when starting a subscription or adjusting future orders on an active subscription.